[FIX] mrp_multi_level: when a product is a kit do not check rule
When you procure a kit, it doesn't matter what the route configuration
is on the product itself but on its children, so it is not needed to
check anything else but the BoM to assign the 'phantom' supply
method.

Also do not assign a BoM when the supply method is not phantom or manufacture.